+++ Js2IL.Tests/Async/ExecutionTests.cs
@@ -59,5 +59,8 @@ namespace Js2IL.Tests.Async
 
         [Fact]
         public Task Async_ClassMethod_CallsOtherAsync() { var testName = nameof(Async_ClassMethod_CallsOtherAsync); return ExecutionTest(testName); }
+
+        [Fact]
+        public Task Async_Inheritance_SuperAsyncMethod() { var testName = nameof(Async_Inheritance_SuperAsyncMethod); return ExecutionTest(testName); }
     }
 }