+++ Js2IL.Tests/Classes/JavaScript/Classes_Inheritance_SuperConstructor_Args.js
@@ -1,0 +1,13 @@
+class B {
+  constructor(x) {
+    this.x = x;
+  }
+}
+
+class D extends B {
+  constructor() {
+    super(5);
+  }
+}
+
+console.log(new D().x);